+++ src/Aks/Aks/Commands/NewAzureRmAks.cs
@@ -140,7 +140,7 @@ namespace Microsoft.Azure.Commands.Aks
         [Parameter(Mandatory = false, HelpMessage = "Whether to enable host based OS and data drive")]
         public SwitchParameter EnableEncryptionAtHost { get; set; }
 
-        [Parameter(Mandatory = false, HelpMessage = "whether to enable UltraSSD")]
+        [Parameter(Mandatory = false, HelpMessage = "Whether to enable UltraSSD")]
         public SwitchParameter EnableUltraSSD { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "The OS configuration of Linux agent nodes.")]
@@ -154,6 +154,24 @@ namespace Microsoft.Azure.Commands.Aks
 
         [Parameter(Mandatory = false, HelpMessage = "The ID for Proximity Placement Group.")]
         public string PPG { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "Whether to use a FIPS-enabled OS.")]
+        public SwitchParameter EnableFIPS { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "The parameters to be applied to the cluster-autoscaler.")]
+        public ManagedClusterPropertiesAutoScalerProfile AutoScalerProfile { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "The GpuInstanceProfile to be used to specify GPU MIG instance profile for supported GPU VM SKU.")]
+        [PSArgumentCompleter("MIG1g", "MIG2g", "MIG3g", "MIG4g", "MIG7g")]
+        public string GpuInstanceProfile { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "Whether to use use Uptime SLA.")]
+        public SwitchParameter EnableUptimeSLA { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "The name of the Edge Zone.")]
+        public string EdgeZone { get; set; }
 
         private AcsServicePrincipal acsServicePrincipal;
 
@@ -387,6 +405,25 @@ namespace Microsoft.Azure.Commands.Aks
             //{
             //    managedCluster.EnablePodSecurityPolicy = EnablePodSecurityPolicy;
             //}
+            if (this.IsParameterBound(c => c.AutoScalerProfile))
+            {
+                managedCluster.AutoScalerProfile = AutoScalerProfile;
+            }
+            if (this.IsParameterBound(c => c.EnableUptimeSLA))
+            {
+                if (EnableUptimeSLA.ToBool())
+                {
+                    managedCluster.Sku = new ManagedClusterSKU(name: "Basic", tier: "Paid");
+                }
+                else
+                {
+                    managedCluster.Sku = new ManagedClusterSKU(name: "Basic", tier: "Free");
+                }
+            }
+            if (this.IsParameterBound(c => c.EdgeZone))
+            {
+                managedCluster.ExtendedLocation = new ExtendedLocation(name: EdgeZone, type: "EdgeZone");
+            }
 
             return managedCluster;
         }
@@ -525,6 +562,14 @@ namespace Microsoft.Azure.Commands.Aks
             if (this.IsParameterBound(c => c.PPG))
             {
                 defaultAgentPoolProfile.ProximityPlacementGroupID = PPG;
+            }
+            if (EnableFIPS.IsPresent)
+            {
+                defaultAgentPoolProfile.EnableFIPS = EnableFIPS.ToBool(); 
+            }
+            if (this.IsParameterBound(c => c.GpuInstanceProfile))
+            {
+                defaultAgentPoolProfile.GpuInstanceProfile = GpuInstanceProfile;
             }
 
             defaultAgentPoolProfile.Mode = NodePoolMode;