+++ src/Aks/Aks/Commands/NewAzureRmAksNodePool.cs
@@ -119,6 +119,15 @@ namespace Microsoft.Azure.Commands.Aks
         [Parameter(Mandatory = false, HelpMessage = "The max price (in US Dollars) you are willing to pay for spot instances. Possible values are any decimal value greater than zero or -1 which indicates default price to be up-to on-demand.")]
         public double? SpotMaxPrice { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Whether to use a FIPS-enabled OS")]
+        public SwitchParameter EnableFIPS { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "The GpuInstanceProfile to be used to specify GPU MIG instance profile for supported GPU VM SKU.")]
+        [PSArgumentCompleter("MIG1g", "MIG2g", "MIG3g", "MIG4g", "MIG7g")]
+        public string GpuInstanceProfile { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -263,6 +272,14 @@ namespace Microsoft.Azure.Commands.Aks
             if (this.IsParameterBound(c => c.SpotMaxPrice))
             {
                 agentPool.SpotMaxPrice = SpotMaxPrice;
+            }
+            if (EnableFIPS.IsPresent)
+            {
+                agentPool.EnableFIPS = EnableFIPS.ToBool();
+            }
+            if (this.IsParameterBound(c => c.GpuInstanceProfile))
+            {
+                agentPool.GpuInstanceProfile = GpuInstanceProfile;
             }
 
             return agentPool;