+++ src/App/custom/New-AzContainerAppSecretObject.ps1
@@ -23,7 +23,7 @@ Create an in-memory object for Secret.
 .Outputs
 Microsoft.Azure.PowerShell.Cmdlets.App.Models.Api20220301.Secret
 .Link
-https://docs.microsoft.com/powershell/module/az./new-azcontainerappsecretobject
+https://learn.microsoft.com/powershell/module/az./new-azcontainerappsecretobject
 #>
 function New-AzContainerAppSecretObject {
     [OutputType('Microsoft.Azure.PowerShell.Cmdlets.App.Models.Api20220301.Secret')]