+++ src/Marketplace/Marketplace.Autorest/exports/Get-AzMarketplaceBillingPrivateStoreAccount.ps1
@@ -43,7 +43,7 @@ INPUTOBJECT <IMarketplaceIdentity>: Identity Parameter
   [PrivateStoreId <String>]: The store ID - must use the tenant ID
   [RequestApprovalId <String>]: The request approval ID to get create or update
 .Link
-https://docs.microsoft.com/powershell/module/az.marketplace/get-azmarketplacebillingprivatestoreaccount
+https://learn.microsoft.com/powershell/module/az.marketplace/get-azmarketplacebillingprivatestoreaccount
 #>
 function Get-AzMarketplaceBillingPrivateStoreAccount {
 [OutputType([System.String])]