+++ src/Synapse/Synapse.Autorest/exports/Remove-AzSynapseKustoPoolAttachedDatabaseConfiguration.ps1
@@ -43,7 +43,7 @@ INPUTOBJECT <ISynapseIdentity>: Identity Parameter
   [SubscriptionId <String>]: The ID of the target subscription.
   [WorkspaceName <String>]: The name of the workspace
 .Link
-https://docs.microsoft.com/powershell/module/az.synapse/remove-azsynapsekustopoolattacheddatabaseconfiguration
+https://learn.microsoft.com/powershell/module/az.synapse/remove-azsynapsekustopoolattacheddatabaseconfiguration
 #>
 function Remove-AzSynapseKustoPoolAttachedDatabaseConfiguration {
 [OutputType([System.Boolean])]