+++ src/Network/Network/ChangeLog.md
@@ -20,8 +20,15 @@
 
 ## Upcoming Release
 * Fixed the incorrect type of `-TotalBytesPerSession` in `New-AzNetworkWatcherPacketCapture`
+
+## Version 5.3.0
+* Added samples for retrieving Private Link IP Configuration using 'New-AzApplicationGatewayPrivateLinkIpConfiguration' with fix [#20440]
+* Added `DdosProtectionPlan` property in `AzPublicIpAddress`
+* Updated mapping in `AzPublicIpAddress` to always show/create DdosSettings
 * Fixed a bug that added Ddos related properties when viewing PublicIpAddress and DdosProtectionPlan objects
-* Fixed a Bug for Set-AzIpGroup cmdlet to support the -whatif parameter
+* Fixed a Bug for Set-AzIpGroup cmdlet to support the `-WhatIf` parameter
+* Fixed a Bug for `Add-AzLoadBalancerFrontendIpConfig`, `Add-AzLoadBalancerProbeConfig`, `Add-AzLoadBalancerBackendAddressPoolConfig`, `Set-AzLoadBalancer`, `New-AzLoadBalancerRuleConfig`, `Remove-AzLoadBalancerInboundNatRuleConfig` cmdlets to support the `-WhatIf` parameter. [#20416]
+* Fixed a bug for DestinationPortBehavior in `Get-AzNetworkWatcherConnectionMonitor`, `New-AzNetworkWatcherConnectionMonitor` powershell command by adding this properties to get and set the DestinationPortBehavior information. [#15996]
 
 ## Version 5.2.0
 * Added optional parameters `CustomBlockResponseStatusCode` and `CustomBlockResponseBody` parameter to `AzApplicationGatewayFirewallPolicySettings`